Refactor create_student
