+++ lib/create_books.rb
